:poop: 4

+++ src/control/ControlData.java
@@ -13,8 +13,8 @@ import java.awt.Point;
  */
 public class ControlData {
     
-    private final double R = 100;
-    public double x, y;
+    private final float R = 100;
+    public float x, y;
     
     // input `p` is mouse - center
     public ControlData(Point p) {
@@ -24,8 +24,8 @@ public class ControlData {
             x = p.x / R;
             y = p.y / R;
         } else {
-            x = p.x / d;
-            y = p.y / d;
+            x = (float) (p.x / d);
+            y = (float) (p.y / d);
         }
     }
 }
